# control-plane/api/openapi.yaml
openapi: 3.0.3
info:
  title: session-platform control plane API
  version: 0.1.0
  description: >
    REST surface the SPA uses to create, list and operate sessions. Each
    session maps 1:1 to a dedicated data plane pod (AC-A2). Non-active access is
    uniform "ensure active then serve": read, write and switch promote
    idle->active (AC-C1) or restore snapshot->active (AC-B2) before serving.
    A session picks a workload type at creation (AC-E1): `shell` (the default)
    or `claude-code`. The type is immutable for the session's lifetime — no
    endpoint changes it, and a restore keeps it — so switching workloads means
    creating a new session. For `shell`, write injects the payload into the PTY
    stdin (AC-D2); for `claude-code`, write asynchronously queues one prompt
    execution (AC-E2). Both read paths return workload output after an offset
    cursor, with offset=0 returning the full history (AC-D3/E3). Claude Code
    credentials are projected from a platform Secret only into a localhost
    credential-proxy sidecar; the tool-running container receives no real
    token. Its optional session model is immutable and defaults to
    `platform-default` (AC-E6).
servers:
  - url: /api/v1
paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
  /sessions:
    get:
      summary: List sessions
      operationId: listSessions
      responses:
        "200":
          description: All sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
    post:
      summary: Create a session (provisions a dedicated pod, goes active)
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSessionRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Malformed body, unknown field, trailing JSON, or invalid name, workload type, or model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
  /sessions/{id}:
    get:
      summary: Get a single session
      operationId: getSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: The session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /sessions/{id}/read:
    post:
      summary: Read workload output since an offset cursor (state-branched, AC-C2)
      operationId: readSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReadRequest" }
      responses:
        "200":
          description: Read result with the branch that served it
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReadResult" }
        "400":
          description: Malformed request body, unknown field, trailing JSON, or invalid offset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
  /sessions/{id}/write:
    post:
      summary: Write to a session (state-branched, AC-C3)
      operationId: writeSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriteRequest" }
      responses:
        "200":
          description: Write result with the branch that served it
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "400":
          description: Malformed request body, unknown field, or trailing JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
        "429":
          description: The workload prompt queue is full; retry after an accepted prompt completes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "507":
          description: The workload output quota is full; existing output remains readable and restorable but new writes are refused
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /sessions/{id}/switch:
    post:
      summary: Switch to a session, restoring from snapshot if needed (AC-C4)
      operationId: switchSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: The now-active session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Malformed body, trailing JSON, or a non-empty request object
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
        "409":
          description: Concurrent state change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  responses:
    RequestEntityTooLarge:
      description: The JSON wire body exceeds 8 MiB, or a Claude Code prompt exceeds the 1 MiB decoded payload limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    State:
      type: string
      enum: [active, idle, snapshot]
    WorkloadType:
      type: string
      enum: [shell, claude-code]
      default: shell
      description: >
        Which data plane workload the session runs (AC-E1). Fixed at creation
        and immutable afterwards. `shell` is an interactive PTY shell
        (AC-D1~D5); `claude-code` is the Claude Code CLI workload (AC-E2~E6),
        which requires the control plane to have an image configured for it.
    Checkpoint:
      type: object
      properties:
        ref: { type: string }
        sizeBytes: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        reclaimed: { type: string }
    Session:
      type: object
      required: [id, name, state, workloadType, createdAt, lastAccess]
      properties:
        id: { type: string }
        name: { type: string }
        workloadType: { $ref: "#/components/schemas/WorkloadType" }
        state: { $ref: "#/components/schemas/State" }
        model:
          type: string
          description: >
            Immutable Claude Code model identifier. Present only for
            `claude-code`; omitted at creation resolves to `platform-default`.
        pod: { type: string }
        createdAt: { type: string, format: date-time }
        lastAccess: { type: string, format: date-time }
        checkpoint: { $ref: "#/components/schemas/Checkpoint" }
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          description: Leading and trailing whitespace is trimmed; at least one non-whitespace character is required.
        workloadType:
          allOf:
            - $ref: "#/components/schemas/WorkloadType"
          description: >
            Optional. Omitted means `shell`, so a request that predates the
            workload type axis behaves unchanged. A value outside the enum is
            rejected with 400 before any pod is provisioned.
        model:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$'
          description: >
            Claude Code model for this session. Valid only when workloadType is
            `claude-code`; omitted resolves to `platform-default`. Credentials
            are never accepted in this request.
    WriteRequest:
      type: object
      additionalProperties: false
      properties:
        payload:
          type: string
          description: >
            Raw workload input. Claude Code writes are limited to 1 MiB of UTF-8
            bytes per prompt; larger writes return 413.
    ReadRequest:
      type: object
      additionalProperties: false
      properties:
        offset:
          type: integer
          format: int64
          default: 0
          minimum: 0
          description: >
            Cursor issued as nextOffset by the previous read. 0 (or omitted)
            returns the full output accumulated since session start (AC-D3);
            an offset past the current length yields an empty payload.
    ReadResult:
      type: object
      properties:
        session: { $ref: "#/components/schemas/Session" }
        path:
          type: string
          description: which state branch served the read
        payload:
          type: string
          description: workload stdout/stderr accumulated after the requested offset
        nextOffset:
          type: integer
          format: int64
          description: >
            cursor to pass as offset on the next read to receive only new
            output (the current accumulated length)
    WriteResult:
      type: object
      properties:
        session: { $ref: "#/components/schemas/Session" }
        path:
          type: string
          description: which state branch served the write
    Error:
      type: object
      properties:
        error: { type: string }